pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aigis"
version = "0.1.0"
description = "AI Execution Governance Linter — static analysis for unsafe AI autonomy"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.scripts]
aeg = "aigis.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/aigis"]

[tool.pytest.ini_options]
testpaths = ["tests"]
